feat(HotelMonitoring): added some minor changes for preparing writing funnels

+++ src/main/java/com/greenfox/kryptonite/projectx/model/funnels/StepAttributes.java
@@ -11,9 +11,9 @@ import lombok.ToString;
 @NoArgsConstructor
 @ToString
 public class StepAttributes extends DataAttributes{
-  private long percent;
+  private int percent;
 
-  public StepAttributes(String path, int count, Long percent) {
+  public StepAttributes(String path, int count, int percent) {
     super(path, count);
     this.percent = percent;
   }
